+++ Network/Telegram/API/Bot/Object.hs
@@ -4,7 +4,6 @@ import Network.Telegram.API.Bot.Object.Update as Exports
 import Network.Telegram.API.Bot.Object.Moving as Exports
 import Network.Telegram.API.Bot.Object.Message as Exports
 import Network.Telegram.API.Bot.Object.Member as Exports
-import Network.Telegram.API.Bot.Object.Keyboard as Exports
 import Network.Telegram.API.Bot.Object.Callback as Exports
 
 import "base" Data.Kind (Constraint)